requirements.txt
asgiref==3.5.1
asttokens==2.0.5
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dj-static==0.0.6
Django==4.0.4
django-bootstrap4==22.1
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-location-field==2.1.0
django-mapbox-location-field==2.0.0
django-multiselectfield==0.1.12
djangorestframework==3.13.1
executing==0.8.3
flake8==4.0.1
icecream==2.1.2
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
pkg_resources==0.0.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.3.0
pyotp==2.6.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
static3==0.7.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
